Refactor FantasyPlayer to use Store
* Move Repo actions into FantasyPlayer.Store
* Closes #285
