+++ lib/ex338_web/admin/championship_result.ex
@@ -10,7 +10,7 @@ defmodule Ex338Web.ExAdmin.ChampionshipResult do
         input championship_result, :championship,
           collection: Ex338.Championship.all()
         input championship_result, :fantasy_player,
-          collection: Ex338.FantasyPlayer.get_all_players()
+          collection: Ex338.FantasyPlayer.Store.get_all_players()
         input championship_result, :rank
         input championship_result, :points
       end
