+++ lib/ex338_web/admin/injured_reserve.ex
@@ -12,11 +12,11 @@ defmodule Ex338Web.ExAdmin.InjuredReserve do
         input waiver, :fantasy_team,
           collection: Ex338.Repo.all(FantasyTeam.alphabetical(FantasyTeam))
         input waiver, :add_player,
-          collection: Ex338.FantasyPlayer.get_all_players()
+          collection: Ex338.FantasyPlayer.Store.get_all_players()
         input waiver, :remove_player,
-          collection: Ex338.FantasyPlayer.get_all_players()
+          collection: Ex338.FantasyPlayer.Store.get_all_players()
         input waiver, :replacement_player,
-          collection: Ex338.FantasyPlayer.get_all_players()
+          collection: Ex338.FantasyPlayer.Store.get_all_players()
         input waiver, :status, collection: Ex338.InjuredReserve.status_options()
       end
     end
