+++ test/ex338/fantasy_player_repo_test.exs
@@ -16,17 +16,6 @@ defmodule Ex338.FantasyPlayerRepoTest do
     end
   end
 
-  describe "player_with_sport!/2" do
-    test "returns a player with the sports league preloaded" do
-      sport = insert(:sports_league)
-      player = insert(:fantasy_player, sports_league: sport)
-
-      result = FantasyPlayer.player_with_sport!(FantasyPlayer, player.id)
-
-      assert result.sports_league.id == sport.id
-    end
-  end
-
   describe "names_and_ids/1" do
     test "selects names and ids" do
       player_a = insert(:fantasy_player, player_name: "A")
@@ -146,44 +135,6 @@ defmodule Ex338.FantasyPlayerRepoTest do
     end
   end
 
-  describe "get_avail_players_for_champ/2" do
-    test "returns unowned players in a league for a championship" do
-      league = insert(:fantasy_league)
-      league_b = insert(:fantasy_league)
-      team = insert(:fantasy_team, team_name: "Brown", fantasy_league: league)
-      team_b = insert(:fantasy_team, team_name: "Axel", fantasy_league: league_b)
-
-      sport = insert(:sports_league)
-      other_sport = insert(:sports_league)
-
-      drafted_player =
-        insert(:fantasy_player, player_name: "E", draft_pick: false,
-          sports_league: sport)
-      insert(:roster_position, fantasy_team: team, fantasy_player: drafted_player)
-      avail_player =
-        insert(:fantasy_player, player_name: "D", draft_pick: false,
-          sports_league: sport)
-      insert(:roster_position, fantasy_team: team_b, fantasy_player: avail_player)
-      unowned_player =
-        insert(:fantasy_player, player_name: "C", draft_pick: false,
-          sports_league: sport)
-      _pick_player =
-        insert(:fantasy_player, player_name: "B", draft_pick: true,
-          sports_league: sport)
-      _other_sport_player =
-        insert(:fantasy_player, player_name: "A", draft_pick: false,
-          sports_league: other_sport)
-
-      result = FantasyPlayer.get_avail_players_for_champ(league.id, sport.id)
-
-      [result_c, result_d] = result
-
-      assert Enum.count(result) == 2
-      assert result_c.id == unowned_player.id
-      assert result_d.id == avail_player.id
-    end
-  end
-
   describe "avail_players_for_champ/1" do
     test "query for unowned players in a league for a championship" do
       league = insert(:fantasy_league)
@@ -222,44 +173,6 @@ defmodule Ex338.FantasyPlayerRepoTest do
       assert Enum.count(result) == 2
       assert result_c.id == unowned_player.id
       assert result_d.id == avail_player.id
-    end
-  end
-
-  describe "get_next_championship/3" do
-    test "returns the next championship for a player" do
-      league = insert(:fantasy_league, year: 2017)
-      sport = insert(:sports_league)
-      other_sport = insert(:sports_league)
-      insert(:league_sport, fantasy_league: league, sports_league: sport)
-      insert(:league_sport, fantasy_league: league, sports_league: other_sport)
-
-      _prev_event = insert(:championship, sports_league: sport,
-        championship_at: CalendarAssistant.days_from_now(-5), year: 2017)
-      _other_event = insert(:championship, sports_league: other_sport,
-        championship_at: CalendarAssistant.days_from_now(10), year: 2017)
-      event = insert(:championship, sports_league: sport,
-        championship_at: CalendarAssistant.days_from_now(14), year: 2017)
-      player = insert(:fantasy_player, sports_league: sport)
-
-      result =
-        FantasyPlayer.get_next_championship(FantasyPlayer, player.id, league.id)
-
-      assert result.championship_at == event.championship_at
-    end
-
-    test "ignores championships next year" do
-      league = insert(:fantasy_league, year: 2017)
-      sport = insert(:sports_league)
-      insert(:league_sport, fantasy_league: league, sports_league: sport)
-
-      insert(:championship, sports_league: sport,
-        championship_at: CalendarAssistant.days_from_now(214), year: 2018)
-      player = insert(:fantasy_player, sports_league: sport)
-
-      result =
-        FantasyPlayer.get_next_championship(FantasyPlayer, player.id, league.id)
-
-      assert result == nil
     end
   end
